move deserialization tests to tests_deserialization
